refactor: move field selector to common
